Create missing collections and retry sync on error
Workaround for https://github.com/dotCMS/core/issues/20300

Presently when gulp-webdav-sync attempts to PUT a new file in a
collection which does not yet exist dotCMS tries and fails to create any
missing intermediate collections. For example, if we try to sync
"/foo/bar/baz.txt" and "/foo/bar" does not yet exist then dotCMS will
explode and respond with a 500 error. To work around this bug we handle
errors by checking whether the target file's parent collection is
missing and if so create it and retry the sync.